backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
python-multipart
werkzeug
pydantic[email]
httpx
shapely
python-dotenv
